introduce FileObject and RemoteObject to replace dicts
We need to do a better job keeping track of the local and remote
lists.  We have been storing bits in dicts, which works, but isn't
very object-like.  We do the same operations on those (e.g. stat a
file, read the file to calculate a checksum) many many times, because
we would "forget" as we passed parts of the dicts around.

This patch introduces the new FileObject and RemoteObject classes,
which replace the corresponding dicts returned from fetch_local_list()
and fetch_remote_list().  These cache operations such that they only
happen once per local file or remote object.

This is the first part of the cleanup, which basically just replaces
the dicts with these classes.  Future cleanup will make sure we're not
doing unnecessary redundant work.