Fixes for python 2.4
Earlier commits broke python 2.4 compatiblity.  This begins to repair that,
however it is not complete.  There is still a failure with some of the sha256
operations due to python 2.4 hashlib and hmac incompatibility that I
haven't been able to sort out.

The point of this commit and branch is to let people who care about
python 2.4 have a starting point to finishing fixing it.  Until then,
we will declare that only python 2.6 and above are valid.  N.B. that
RHEL5 with default python 2.4 was released in March 2007, nearly 8 years
ago, and that there do exist python26 packages for RHEL5 that would
let s3cmd work without requiring python 2.4.