Push getting md5 from metadata down into compare
On sync, we were getting the md5 from the object's metadata if the
ETag indicated it was stored via multipart upload, which by default
kicks in at 15MB.  If your tree has lots of >15MB files, you would see
sync call object_info() for each getting the md5.  Even if the size
(obtained from the bucket list) mismatched. :-(

This patch moves the call to _get_object_attrs() down into the compare
function between the src and dst lists, where the sizes already match,
and only if we have md5 checks turned on, and only if a remote file
has a '-' in its md5 (multipart upload). This should reduce the number
of object_info (aka HTTP HEAD) calls, at least for trees with <15MB
files.

To avoid the extra HEAD calls all together, disable md5 checking with
--no-check-md5.